+++ stroom-query-audit/src/main/java/stroom/query/audit/CriteriaStore.java
@@ -3,8 +3,6 @@ package stroom.query.audit;
 import stroom.mapreduce.v2.UnsafePairQueue;
 import stroom.query.api.v2.TableSettings;
 import stroom.query.common.v2.CompiledSorter;
-import stroom.query.common.v2.CompletionListener;
-import stroom.query.common.v2.Coprocessor;
 import stroom.query.common.v2.CoprocessorSettingsMap;
 import stroom.query.common.v2.Data;
 import stroom.query.common.v2.Item;
@@ -16,13 +14,10 @@ import stroom.query.common.v2.StoreSize;
 import stroom.query.common.v2.TableCoprocessorSettings;
 import stroom.query.common.v2.TablePayload;
 
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
-import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.TimeUnit;
 
 /**
  * Used to store the results from a query made on a {@link stroom.query.audit.service.QueryService}
@@ -30,37 +25,40 @@ import java.util.concurrent.atomic.AtomicBoolean;
 public class CriteriaStore implements Store {
 
     private final CoprocessorSettingsMap coprocessorSettingsMap;
-    private final Map<CoprocessorSettingsMap.CoprocessorKey, Coprocessor> coprocessorMap;
     private final Map<CoprocessorSettingsMap.CoprocessorKey, Payload> payloadMap;
 
     private final List<Integer> defaultMaxResultsSizes;
     private final StoreSize storeSize;
-    private final List<CompletionListener> completionListeners = Collections.synchronizedList(new ArrayList<>());
-
-    //results are currently assembled synchronously in getData so the store is always complete
-    private final AtomicBoolean isComplete = new AtomicBoolean(true);
 
     public CriteriaStore(final List<Integer> defaultMaxResultsSizes,
                          final StoreSize storeSize,
                          final CoprocessorSettingsMap coprocessorSettingsMap,
-                         final Map<CoprocessorSettingsMap.CoprocessorKey, Coprocessor> coprocessorMap,
                          final Map<CoprocessorSettingsMap.CoprocessorKey, Payload> payloadMap) {
-
         this.defaultMaxResultsSizes = defaultMaxResultsSizes;
         this.storeSize = storeSize;
         this.coprocessorSettingsMap = coprocessorSettingsMap;
-        this.coprocessorMap = coprocessorMap;
         this.payloadMap = payloadMap;
     }
 
     @Override
     public void destroy() {
-        //nothing to do as this store doesn't hold any query state
+        // Do nothing to do as this store doesn't hold any query state
     }
 
     @Override
     public boolean isComplete() {
-        return isComplete.get();
+        // Results are currently assembled synchronously in getData so the store is always complete
+        return true;
+    }
+
+    @Override
+    public void awaitCompletion() {
+    }
+
+    @Override
+    public boolean awaitCompletion(final long timeout, final TimeUnit unit) {
+        // Results are currently assembled synchronously in getData so the store is always complete
+        return true;
     }
 
     @Override
@@ -109,16 +107,5 @@ public class CriteriaStore implements Store {
     @Override
     public StoreSize getStoreSize() {
         return storeSize;
-    }
-
-    @Override
-    public void registerCompletionListener(final CompletionListener completionListener) {
-        if (isComplete.get()) {
-            //immediate notification
-            completionListener.onCompletion();
-        } else {
-            //TODO this is currently of no use but when incremenntal queries are implemented it will be needed
-            completionListeners.add(Objects.requireNonNull(completionListener));
-        }
     }
 }